requirements.txt
fastapi
uvicorn[standard]
numpy
pillow
opencv-python
torch
torchvision
diffusers
transformers
accelerate
safetensors
huggingface_hub
